put files to remove on folder
