+++ packages/oc-template-typescript-react-compiler/lib/compileServer.js
@@ -1,34 +1,25 @@
-'use strict';
-
-const async = require('async');
+const vite = require('vite');
 const fs = require('fs-extra');
-const hashBuilder = require('oc-hash-builder');
-const MemoryFS = require('memory-fs');
 const path = require('path');
-
-const {
-  compiler,
-  configurator: { server: webpackConfigurator }
-} = require('./to-abstract-base-template-utils/oc-webpack');
+const { callbackify } = require('util');
+const hashBuilder = require('oc-hash-builder');
 const higherOrderServerTemplate = require('./higherOrderServerTemplate');
 
-const hasTsExtension = (file) => !!file.match(/\.tsx?$/);
+const externals = ['got']
 
-module.exports = (options, callback) => {
+async function compileServer(options) {
   const componentPath = options.componentPath;
   const serverFileName = options.componentPackage.oc.files.data;
-  const usingTypescript = hasTsExtension(serverFileName);
   let serverPath = path.join(options.componentPath, serverFileName);
   if (process.platform === 'win32') {
     serverPath = serverPath.split('\\').join('\\\\');
   }
   const publishFileName = options.publishFileName || 'server.js';
   const publishPath = options.publishPath;
-  const stats = options.verbose ? 'verbose' : 'errors-only';
-  const dependencies = options.componentPackage.dependencies || {};
+  // const dependencies = options.componentPackage.dependencies || {};
   const componentName = options.componentPackage.name;
   const componentVersion = options.componentPackage.version;
-  const production = options.production;
+  const production = !!options.production;
 
   const higherOrderServerContent = higherOrderServerTemplate({
     serverPath,
@@ -38,61 +29,37 @@ module.exports = (options, callback) => {
   const tempFolder = path.join(serverPath, '../_package/temp');
   const higherOrderServerPath = path.join(tempFolder, '__oc_higherOrderServer.ts');
 
-  const config = webpackConfigurator({
-    componentPath,
-    serverPath: higherOrderServerPath,
-    usingTypescript,
-    publishFileName,
-    dependencies,
-    stats,
-    production
-  });
+  try {
+    await fs.outputFile(higherOrderServerPath, higherOrderServerContent);
 
-  async.waterfall(
-    [
-      (next) => fs.outputFile(higherOrderServerPath, higherOrderServerContent, next),
-      (next) => {
-        return compiler(config, next);
-      },
-      (data, next) => {
-        const basePath = path.join(tempFolder, 'build');
-        const memory = new MemoryFS(data);
-        const getCompiled = (fileName) => memory.readFileSync(`${basePath}/${fileName}`, 'UTF8');
-
-        return fs.ensureDir(publishPath, (err) => {
-          if (err) return next(err);
-          const result = { 'server.js': getCompiled(config.output.filename) };
-
-          if (!production) {
-            try {
-              result['server.js.map'] = getCompiled(`${config.output.filename}.map`);
-            } catch (e) {
-              // skip sourcemap if it doesn't exist
-            }
-          }
-
-          next(null, result);
-        });
-      },
-      (compiledFiles, next) => {
-        return async.eachOf(
-          compiledFiles,
-          (fileContent, fileName, next) =>
-            fs.writeFile(path.join(publishPath, fileName), fileContent, next),
-          (err) =>
-            next(
-              err,
-              err
-                ? null
-                : {
-                  type: 'node.js',
-                  hashKey: hashBuilder.fromString(compiledFiles[publishFileName]),
-                  src: publishFileName
-                }
-            )
-        );
+    const result = await vite.build({
+      root: componentPath,
+      mode: production ? 'production' : 'development',
+      logLevel: 'silent',
+      build: {
+        lib: { entry: higherOrderServerPath, formats: ['cjs'] },
+        write: false,
+        minify: production,
+        rollupOptions: {
+          external: externals,
+        }
       }
-    ],
-    (err, data) => fs.remove(tempFolder, (_err2) => callback(err, data))
-  );
-};
+    });
+    const out = Array.isArray(result) ? result[0] : result;
+    const bundle = out.output[0].code;
+
+    await fs.ensureDir(publishPath);
+    await fs.writeFile(path.join(publishPath, publishFileName), bundle);
+
+    return {
+      type: 'node.js',
+      hashKey: hashBuilder.fromString(bundle),
+      src: publishFileName
+    }
+  } finally {
+    await fs.remove(tempFolder);
+  }
+}
+
+
+module.exports = callbackify(compileServer);
