+++ packages/oc-template-typescript-react-compiler/lib/to-remove/to-abstract-base-template-utils/oc-webpack/lib/configurator/client/index.js
